Added an intelligent filename parser, supoprting classes and tests. Removed the code from AppWindow where i was manually setting data.
git-svn-id: https://tanuki2.googlecode.com/svn/trunk@26 3774bd98-f447-0410-8733-1d6108554824

+++ src/golly/tanuki2/qa/DataTest.java
@@ -35,10 +35,10 @@ public class DataTest extends Assert {
 		assertTrue(ad1.equals(ad2));
 		assertTrue(ad1.hashCode() == ad2.hashCode());
 
-		ad2.setYear(null);
+		ad2.setYear((Integer)null);
 		assertFalse(ad1.equals(ad2));
 		assertTrue(ad1.hashCode() != ad2.hashCode());
-		ad1.setYear(null);
+		ad1.setYear((Integer)null);
 		assertTrue(ad1.equals(ad2));
 		assertTrue(ad1.hashCode() == ad2.hashCode());
